# backend/api/swagger.yaml
openapi: "3.0.0"
info:
  title: API Music Room Suricates 
  description: This is a MusicRoom server with all the requests available to the users.
    You can authentify yourself to get access to the requests.
  version: 1.0.0
servers:
  - url: "http://62.34.5.191:45559/api"
  - url: "http://localhost:45559/api"
  - url: "http://localhost:46000/api"
...
paths:
...
  /search:
    get:
      tags:
      - "search"
      summary: "Gives the result of the research"
      operationId: "searchGlobal"
      parameters:
        - name: "value"
          in: "query"
          description: "value of the research"
          required: false
          schema:
            type: "string"
        - name: "type"
          in: "query"
          description: "type to search in ( pseudos, playlists, events )"
          required: true
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "The max number of elements returned"
          schema:
            type: "integer"
        - name: "musicalStyle"
          in: "query"
          description: "Musical style of the research"
          required: false
          schema:
            type: "string"
      responses:
        200: 
          description: |
            `Success - Return the result of the search !`
        400:
          description: |
            __{code:__ '0'__}__ - `Wrong parameter!`

            __{code:__ '2'__}__ - `Type must be specified !`

            __{code:__ '3'__}__ - `Limit must be an integer !`

            __{code:__ '4'__}__ - `This type doesn't exist !`
        500: 
          description: |
           __{code:__ '0'__}__ - `Server error or Unhandled error!`
      security: []
  /users/{userId}/searchFriends:
    get:
      tags:
      - "search"
      summary: "Gives the result of friend research"
      operationId: "searchFriends"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
        - name: "value"
          in: "query"
          description: "value of the research"
          required: true
          schema:
            type: "string"
      responses:
        200: 
          description: |
            `Success - Return the result of the search !`
        400:
          description: |
            __{code:__ '0'__}__ - `Wrong parameters!`
        500: 
          description: |
           __{code:__ '0'__}__ - `Server error or Unhandled error!`
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []